# docs/swagger.yaml
definitions:
...
  CreateRemodelVO:
    properties:
      content:
        description: 改造内容
        maxLength: 100
        type: string
      director:
        description: 责任人
        type: string
      location:
        description: 改造地点
        type: string
      moldCode:
        description: 模具编码
        type: string
      remodelEndTime:
        description: 改造结束时间
        type: string
      remodelStartTime:
        description: 改造开始时间
        type: string
      type:
        description: 改造类别
        type: string
    required:
    - director
    - location
    - moldCode
    - remodelEndTime
    - remodelStartTime
    - type
    type: object
...
  DeleteVO:
    properties:
      ids:
        items:
          type: integer
        minItems: 1
        type: array
    type: object
...
  IDVO:
    properties:
      id:
        type: integer
    type: object
...
  PageRemodelOutVO:
    properties:
      code:
        description: 改造编码
        type: string
      content:
        description: 改造内容
        type: string
      delayDay:
        description: 延期天数，范围：1-7
        type: integer
      director:
        description: 责任人
        type: string
      finishTime:
        description: 改造完成时间
        type: string
      id:
        type: integer
      isDelay:
        description: 是否延期
        type: string
      location:
        description: 改造地点
        type: string
      moldCode:
        description: 模具编码
        type: string
      partCodes:
        description: 零件号
        type: string
      projectName:
        description: 项目名称
        type: string
      remodelEndTime:
        description: 改造结束时间
        type: string
      remodelStartTime:
        description: 改造开始时间
        type: string
      status:
        description: 状态
        type: string
      type:
        description: 改造类别
        type: string
      withdrawReason:
        description: 撤销原因
        type: string
    type: object
...
  UpdateRemodelVO:
    properties:
      content:
        description: 改造内容
        type: string
      director:
        description: 责任人
        type: string
      id:
        type: integer
      location:
        description: 改造地点
        type: string
      moldCode:
        description: 模具编码
        type: string
      remodelEndTime:
        description: 改造结束时间
        type: string
      remodelStartTime:
        description: 改造开始时间
        type: string
      type:
        description: 改造类别
        type: string
    type: object
...
  mold.StatusVO:
    properties:
      id:
        description: ID
        type: integer
      reason:
        description: 撤单原因
        type: string
    type: object
...
info:
  contact: {}
  description: 模具管理API
  title: 模具管理API
  version: "1.0"
paths:
...
  /mold/remodel:
    delete:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/DeleteVO'
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 删除模具改造
      tags:
      - 模具改造
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/CreateRemodelVO'
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 添加模具改造
      tags:
      - 模具改造
    put:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/UpdateRemodelVO'
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 更新模具改造
      tags:
      - 模具改造
  /mold/remodel/completed:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/IDVO'
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 完工
      tags:
      - 模具改造
  /mold/remodel/one:
    get:
      consumes:
      - application/json
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/PageRemodelOutVO'
      summary: 模具改造查看
      tags:
      - 模具改造
  /mold/remodel/page:
    get:
      consumes:
      - application/json
      parameters:
      - description: 改造编码
        in: query
        name: code
        type: string
      - description: 改造编码模糊
        in: query
        name: codeLike
        type: string
      - description: 模具编号和模具名称模糊查找，该参数有值时忽略其它条件参数
        in: query
        name: codeOrName
        type: string
      - description: 改造内容
        in: query
        name: content
        type: string
      - description: 当前页
        in: query
        name: currentPage
        type: integer
      - description: 责任人
        in: query
        name: director
        type: string
      - description: 改造完成时间
        in: query
        name: finishTime
        type: string
      - in: query
        name: id
        type: integer
      - description: 是否延期
        in: query
        name: isDelay
        type: string
      - description: 改造地点
        in: query
        name: location
        type: string
      - description: 模具编码
        in: query
        name: moldCode
        type: string
      - description: 改造结束时间Begin
        in: query
        name: remodelEndTimeBegin
        type: string
      - description: 改造结束时间End
        in: query
        name: remodelEndTimeEnd
        type: string
      - description: 改造开始时间Begin
        in: query
        name: remodelStartTimeBegin
        type: string
      - description: 改造开始时间End
        in: query
        name: remodelStartTimeEnd
        type: string
      - description: 每页条数
        in: query
        name: size
        type: integer
      - description: 状态
        in: query
        name: status
        type: string
      - description: 改造类别
        in: query
        name: type
        type: string
      - description: 撤销原因
        in: query
        name: withdrawReason
        type: string
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/PageRemodelOutVO'
      summary: 模具改造分页
      tags:
      - 模具改造
  /mold/remodel/withdraw:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/mold.StatusVO'
      - description: Token
        in: header
        name: AuthToken
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"message":"请求成功","result":"null"}'
          schema:
            type: string
      summary: 撤单
      tags:
      - 模具改造
...
swagger: "2.0"